move injector methods from MinigameServiceImpl to GuiceManagerImpl
